Remove circular dependency
express.d.ts and serve-static.d.ts had a circular dependency, which causes NuGet pain.
Extract core definitions from express.d.ts and move them to express-serve-static-core.d.ts
Update serve-static.d.ts to depend on express-serve-static-core.d.ts
