+++ express-serve-static-core/express-serve-static-core-tests.ts
@@ -1,0 +1,5 @@
+/// <reference path="express-serve-static-core.d.ts" />
+
+import * as express from 'express-serve-static-core';
+
+// null test file - everything should be tested from express.d.ts and serve-static.d.ts
